+++ server.js
@@ -35,15 +35,17 @@ app.use(express.static("public"));
 // Note: Feel free to replace the example routes below with your own
 const usersRoutes = require("./routes/users");
 const widgetsRoutes = require("./routes/widgets");
-const getTask = require('./routes/homePage')
-const editModal = require('./routes/edit-modal')
+const getTasks = require('./routes/homePage')
+const editModal = require('./routes/modal')
+const updateModal = require('./routes/modal-update')
 
 // Mount all resource routes
 // Note: Feel free to replace the example routes below with your own
 app.use("/users", usersRoutes(db));
-app.use('/', getTask(db))
+app.use('/', getTasks(db))
 app.use('/', editModal(db))
-// app.use('/', editModal(db))
+app.use('/', updateModal(db))
+
 // Note: mount other resources here, using the same pattern above
